+++ changelog.md
@@ -14,6 +14,10 @@
 
 - All default IRQ Handler is renamed to `dcd_int_handler()`
 
+### Others
+
+- tud_cdc_n_write_flush() return number of bytes forced to transfer instead of bool 
+
 ## 0.6.0 - 2019.03.30
 
 Added **CONTRIBUTORS.md** to give proper credit for contributors to the stack. Special thanks to [Nathan Conrad](https://github.com/pigrew), [Peter Lawrence](https://github.com/majbthrd) and [William D. Jones](https://github.com/cr1901) and others for spending their precious time to add lots of features and ports for this release.